Refactor widgets into smaller files
