+++ lib/features/sessions/ui/widgets/schedule_view_card.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:fluttercon/common/data/models/models.dart';
+
+class ScheduleViewCard extends StatelessWidget {
+  const ScheduleViewCard({
+    super.key,
+    required this.session,
+    required this.listIndex,
+  });
+
+  final Session session;
+  final int listIndex;
+  @override
+  Widget build(BuildContext context) {
+    return Placeholder();
+  }
+}
